# lib/api-zod/src/generated/types/summarizeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SummarizeBody {
  /** @minLength 1 */
  text: string;
  /** Hint for output language (e.g. 'hi', 'en'). Optional — defaults to matching the input. */
  language?: string;
}