# lib/api-zod/src/generated/types/vocabularyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VocabularyBody {
  /** @minLength 1 */
  word: string;
  /** Optional surrounding sentence for disambiguation */
  context?: string;
}