# lib/api-zod/src/generated/types/vocabularyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VocabularyResponse {
  word: string;
  meaning: string;
  example: string;
  synonyms: string[];
  /** Hindi meaning when the word is in Devanagari */
  meaningHindi?: string;
}